if there is no Global or User Global PLT file create a project local one
This removes the requirement that the user have a plt file in thier
home directory. This change will set the makefile use that PLT file if
its there, but otherwise it will create one local to the project. It
does this transperantly to the user. It also removes the old number 2
option from the README.md. A PLT file is actually not optional, since
its required for the 'make' command to finish. This change should
provide enough information for that to be redundant anyway.