add the PHONY to the makefile for the non-file rules
Rules in make are all geared towards files. It does a lot of stuff to
check and see if files have been updated and need to be rebuilt. When
you have rules that are not files it can result in some wierd
behaviour (as make rebuilds things that should not be rebuilt). You
get around that by telling make what rules are not files with the
'PHONY' variable. You may or may not hit this problem, but having
'PHONY' is just good makefile engineering and avoids those problems
entirely.