Parameterized and renamed quickbuild
Since we are using the script so that it can be called from the erchef
build process, then it does not make sense to keep calling this
'quickbuild'.

The script was also parameterized so that it can drop a binary somewhere
other than priv/solver.

Lastly, the script was written in a way so that it can be called from
anywhere. It will figure out the project directory and compile as
appropriate.